Optimizations + context change
Since 1.19.4 is adding the 'on attacker' subcommand of 'execute', the tag 'ehm.player' is no longer needed. In addition, 'on attacker' allows us to directly find the entity that damaged a player. So searching through entities by ID matching isn't necessary in this case. It's still necessary when trying to trace what entity was hurt by a player, however.

Functions in the function tags are now called in the context of the entity that was damaged. 'on attacker' can be used to efficiently trace the entity the dealt the damage.