Changed the function tag preconditions
The function tags will now execute every time the player hurts an entity or is hurt by an entity. It no longer matters if either have IDs. IDs are not required.

This change was done in anticipation of the addition of Hit Match Lite. Hit Match Lite will offer better performance at the cost of losing features, so it'll only benefit data packs that don't require those features.